# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/get-students": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/student": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "classId": {
                  "example": "any"
                },
                "section": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "attendanceList": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/students/get-attendance": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "classId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "section",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/teachers/get-attendance-by-student": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "studentId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/teacher/checkin": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "example": "any"
                },
                "checkInList": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/teacher/checkout": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "date": {
                  "example": "any"
                },
                "checkOutList": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/teachers/get-attendance": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/teachers/get-attendance-by-teacher": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "teacherId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}